residual resampling for poisson, binomial glmms
Experimental and not well tested for now.  It uses the
simple (not studentized) Pearson residuals on the scale of the response.  This is almost certainly not the best sampling scheme, but seems to give decent
results on simple examples, and I'll need a simulation
study to figure out what would improve it.